requirements.txt
click==8.1.8
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
pendulum==3.0.0
babel==2.17.0
